+++ airwolf/src/shooter.rs
@@ -22,8 +22,8 @@ pub trait Shooter: Entity {
 
         let bprops = self.bullet_props();
         let (x, y) = self.body().draw_pos();
-        let x = x + bprops.x_offset;
-        let y = y + bprops.y_offset;
+        let x = x as f32 + bprops.x_offset;
+        let y = y as f32 + bprops.y_offset;
         let bullet = if self.is_enemy() {
             Bullet::new_enemy(x, y, ctx)
         } else {
